refactor: rewrite to remove objects and use a more data-oriented style
The OOP-like structure doesn't fit Vulkan and makes the code difficult
to read and understand. Instead, all vulkan variables are stored in
`SelectedConfig` inside config.h, and the builders' job is simply to
initialize parts of it.

This approach is not as encapsulated and might not be as "dynamic", but
for the vulkan initialization the encapsulation seems overkill - when
are you going to, for insatnce, be creating queues on-the-fly without
wanting them in plain sight?